requirements.txt
Django==2.0.7
djangorestframework==3.8.2
Faker==0.8.17
pkg-resources==0.0.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
text-unidecode==1.2
